+++ ppx/melange_ppx.ml
@@ -329,7 +329,7 @@ module Obj = struct
             {
               e with
               pexp_desc =
-                Ast_external_mk.record_as_js_object e.pexp_loc label_exprs;
+                Ast_external_mk.record_as_js_object ~loc:e.pexp_loc label_exprs;
             }
         | _ -> Location.raise_errorf ~loc "%%mel.obj requires a record literal"
       in